+++ lib/useTransitionEffect.ts
@@ -3,7 +3,8 @@ import { useMemo, useRef, useEffect, useLayoutEffect } from 'react'
 import {
   cleanSplittedClassName,
   takeTarget,
-  useUniversalEffect
+  useUniversalEffect,
+  onTakeTarget
 } from '@/shared'
 import { useConditionalEffect } from '@/useConditionalEffect'
 import { useIsFirstMountRef } from '@/useIsFirstMountRef'
@@ -40,6 +41,7 @@ type Transition =
   | Enter
   | `${Enter}${From}`
   | `${Enter}${To}`
+  | `${Enter}ed`
   | Leave
   | `${Leave}${From}`
   | `${Leave}${To}`
@@ -49,6 +51,13 @@ type TransitionClassName = Record<Transition, string>
 type UseTransitionEffectOptions = {
   target: Target<HTMLElement | SVGElement>
   show?: boolean
+
+  /**
+   * Whether to apply `display: none;` or `visibility: hidden;` to invisible
+   * If `false`, `display:none;` will be applied
+   * @defaultValue `false`
+   */
+  keepLayout?: boolean
 } & Partial<TransitionClassName>
 
 /**
@@ -80,44 +89,58 @@ const useTransitionEffect: UseEffect<UseTransitionEffectOptions> = ({
   enter,
   enterFrom,
   enterTo,
+  entered,
   leave,
   leaveFrom,
   leaveTo,
   target,
-  show
+  show,
+  keepLayout = false
 }) => {
   const _enter = useCleanClassList(enter)
   const _enterTo = useCleanClassList(enterTo)
   const _enterFrom = useCleanClassList(enterFrom)
+  const _entered = useCleanClassList(entered)
   const _leave = useCleanClassList(leave)
   const _leaveTo = useCleanClassList(leaveTo)
   const _leaveFrom = useCleanClassList(leaveFrom)
+  const classNameRef = useClassListRef(target)
+  const { isFirstMount } = useIsFirstMountRef()
 
   const cleanUp: VFn = () =>
-    [_enter, _enterTo, _enterFrom, _leave, _leaveFrom, _leaveTo].forEach(remove)
+    [
+      _enter,
+      _enterTo,
+      _enterFrom,
+      _entered,
+      _leave,
+      _leaveFrom,
+      _leaveTo
+    ].forEach(removeClassList)
 
-  const { isFirstMount } = useIsFirstMountRef()
-  useConditionalEffect(
-    () => {
-      const ref = takeTarget(target)
-      if (ref) {
-        ref.style.display = 'none'
+  const hide: VFn = () =>
+    onTakeTarget(target, (el) => {
+      if (keepLayout) {
+        el.style.visibility = 'hidden'
+      } else {
+        el.style.display = 'none'
       }
-    },
+    })
+
+  useConditionalEffect(
+    hide,
     [],
     () => isFirstMount && show === false,
     isBrowser ? useLayoutEffect : useEffect
   )
 
-  const classNameRef = useClassListRef(target)
-
-  const add = (classList: string[]): void => {
+  const addClassList = (classList: string[]): void => {
     const originalClass = without(classList, classNameRef.current)
 
     takeTarget(target)?.classList.add(...originalClass)
   }
 
-  const remove = (className: string[]): void => {
+  const removeClassList = (className: string[]): void => {
     const originalClass = without(className, classNameRef.current)
 
     takeTarget(target)?.classList.remove(...originalClass)
@@ -128,42 +151,49 @@ const useTransitionEffect: UseEffect<UseTransitionEffectOptions> = ({
       target,
       entered: show,
       onBeforeEnter: () => {
-        const ref = takeTarget(target)
-        if (ref && ref.style.display === 'none') {
-          ref.style.display = ''
-        }
-        add(_enterFrom)
+        cleanUp()
+        addClassList(_enterFrom)
+
+        onTakeTarget(target, (el) => {
+          if (keepLayout) {
+            if (el.style.visibility === 'hidden') {
+              el.style.visibility = 'visible'
+            }
+          } else {
+            if (el.style.display === 'none') {
+              el.style.display = ''
+            }
+          }
+        })
       },
 
       onEnter: () => {
-        remove(_enterFrom)
-        add(_enterTo)
-        add(_enter)
+        removeClassList(_enterFrom)
+        addClassList(_enterTo)
+        addClassList(_enter)
       },
       onAfterEnter: () => {
-        remove(_enterTo)
-        remove(_enter)
+        removeClassList(_enterTo)
+        removeClassList(_enter)
+        addClassList(_entered)
       },
 
       onBeforeLeave: () => {
         cleanUp()
-        add(_leaveFrom)
+        addClassList(_leaveFrom)
       },
 
       onLeave: () => {
-        remove(_leaveFrom)
-        add(_leaveTo)
-        add(_leave)
+        removeClassList(_leaveFrom)
+        addClassList(_leaveTo)
+        addClassList(_leave)
       },
 
       onAfterLeave: () => {
-        remove(_leaveTo)
-        remove(_leave)
+        removeClassList(_leaveTo)
+        removeClassList(_leave)
 
-        const ref = takeTarget(target)
-        if (ref) {
-          ref.style.display = 'none'
-        }
+        hide()
       }
     },
     []