+++ stories/useTransitionEffect/docs.ts
@@ -73,6 +73,17 @@ const argTypes: ArgTypes = {
     }
   },
 
+  entered: {
+    description: 'Classes to add to the transitioning element once the transition is done. These classes will persist until leave',
+    table: {
+      category,
+      subcategory: subcategory0,
+      type: {
+        summary: 'string | undefined'
+      }
+    }
+  },
+
   leaveFrom: {
     description:
       'Classes to add to the transitioning element before the leave phase starts',
@@ -105,6 +116,22 @@ const argTypes: ArgTypes = {
       subcategory: subcategory0,
       type: {
         summary: 'string | undefined'
+      }
+    }
+  },
+
+  keepLayout: {
+    description:
+      'Whether to apply `display: none;` or `visibility: hidden;` to invisible',
+
+    table: {
+      category,
+      subcategory: subcategory0,
+      type: {
+        summary: 'boolean | undefined'
+      },
+      defaultValue: {
+        summary: 'false'
       }
     }
   },